requirements.txt
a2wsgi==1.10.8
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.1
apache-airflow-core==3.0.1
apache-airflow-providers-common-compat==1.7.0
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.1
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-standard==1.2.0
apache-airflow-task-sdk==1.0.1
appnope==0.1.4
argcomplete==3.6.2
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
azure-core==1.34.0
azure-storage-blob==12.25.1
blinker==1.9.0
cachetools==5.5.2
cadwyn==5.3.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==45.0.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
dnspython==2.7.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.1
fonttools==4.58.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.2.7
kiwisolver==1.4.8
lazy-object-proxy==1.11.0
libcst==1.7.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
narwhals==1.40.0
nest-asyncio==1.6.0
numpy==2.2.6
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pendulum==3.1.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
pyspark==3.5.5
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
retryhttp==1.3.3
rich==14.0.0
rich-argparse==1.7.0
rich-toolkit==0.14.6
rpds-py==0.25.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setproctitle==1.3.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.1
structlog==25.3.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5
traitlets==5.14.3
typer==0.15.4
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.4.0
uuid6==2024.7.10
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.2
xgboost==3.0.1
zipp==3.21.0
